% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data_from_standard_block}
\alias{add_data_from_standard_block}
\title{Title}
\usage{
add_data_from_standard_block(data, block)
}
\arguments{
\item{data}{flow chart data}

\item{block}{code block}
}
\description{
Title
}
