% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{bkill}
\alias{bkill}
\title{Kills any running LSF jobs associated with this Job object}
\usage{
\S4method{bkill}{Job}(.Object, jid = NULL)
}
\description{
Kills any running LSF jobs associated with this Job object
}
\author{
Marcin Imielinski
}

