% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bsub_cmd}
\alias{bsub_cmd}
\title{bsub_cmd}
\usage{
bsub_cmd(cmd, queue, jname = NULL, jlabel = NULL, jgroup = NULL,
  mem = NULL, group = "cgafolk", cwd = NULL, mc.cores = NULL,
  deadline = F)
}
\description{
Makes bsub command that wraps shell command "cmd" to send to queue "queue"
redirecting output / error etc streams to path prefixed by "jname",
optional_args: maximum memory requirements "mem", "jlabel" job label
}
\author{
Marcin Imielinski
}
