% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_nomnoml.R
\name{build_nomnoml_code}
\alias{build_nomnoml_code}
\title{build nomnoml code from data}
\usage{
build_nomnoml_code(data, code, direction = c("down", "right"),
  ranker = c("network-simplex", "tight-tree", "longest-path"),
  arrowSize = 1, edges = c("hard", "rounded"), bendSize = 0.3,
  font = "Calibri", fontSize = 12, lineWidth = 3, padding = 16,
  spacing = 40, leading = 1.25, stroke = "#33322E",
  fill = "#eee8d5", title = "filename", zoom = 1,
  fillArrows = FALSE, acyclicer = "greedy", gutter = 5,
  edgeMargin = 0)
}
\arguments{
\item{data}{A data frame built by `flow_data()`}

\item{code}{Whether to display the code in code blocks or only the header, to be
more compact, if `NA`, the code will be displayed only if no header is
defined by special comments.}

\item{direction}{Whether to orient the chart from top to bottom or left to right}

\item{ranker}{ranker, set to "longest-path" to have all exit points alligned
at the bottom.}

\item{arrowSize}{Arrow size}

\item{edges}{Whether to keep edges `"straight"`, or have them `"rounded"`}

\item{bendSize}{Bend size, if `edges` is `"rounded"`}

\item{font}{Font}

\item{fontSize}{Font size}

\item{lineWidth}{Line Width}

\item{padding}{Padding}

\item{spacing}{Spacing between blocks}

\item{leading}{distance between two baselines of lines of type}

\item{stroke}{Stroke}

\item{fill}{Default filling color}

\item{title}{Title, no effect was observed}

\item{zoom}{Zoom, no effect was observed}

\item{fillArrows}{Whether to fill arrows, no effect was observed}

\item{acyclicer}{Acyclicer, no effect was observed}

\item{gutter}{Gutter, no effect was observed}

\item{edgeMargin}{Edge margin, no effect was observed}
}
\description{
This functions builds the nomnoml code to render from the data built by
`flow_data()`. It's not very useful for the average user to call it directly
but its arguments can be set through `flow_view()` so their definitions are useful.
}
\details{
All arguments, with the exception of `data` and `code`, are
nomnoml directives, enumerated at https://openbase.io/js/nomnoml/documentation .

Some of those don't seem to do anything in the context of this package,
but given the sparsity of nomnoml documentation, we
decided to keep all of them for more flexibility.
}
