% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_C.R
\name{build_plantuml_code_from_c}
\alias{build_plantuml_code_from_c}
\title{Draw Flow Diagram of a C Function}
\usage{
build_plantuml_code_from_c(x, fun = NULL)
}
\arguments{
\item{x}{url of the file}

\item{fun}{function as a string, if missing, \code{build_plantuml_code_from_c()}
will name the possible choices, and return invisibly a named list of the
code of the functions.}
}
\description{
Draw Flow Diagram of a C Function
}
