% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{c}
\alias{c}
\title{Concatenates Job objects built from the same Task}
\usage{
\S4method{c}{Job}(x, ..., recursive = FALSE)
}
\description{
Concatenates Job objects built from the same Task
}
\author{
Marcin Imielinski
}

