% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{cache}
\alias{cache}
\title{caches .rds copy of job object to standard location (TASK.NAME.rds in Job output root directory)}
\usage{
\S4method{cache}{Job}(object, verbose = TRUE)
}
\author{
Marcin Imielinski
}
