% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculateFlowShortScaleFactors.R
\name{CalculateFlowShortScaleFactors}
\alias{CalculateFlowShortScaleFactors}
\title{Calculate the flow short scale factors.}
\usage{
CalculateFlowShortScaleFactors(items)
}
\arguments{
\item{items}{Questionaire items. A numerical vector of the length 16 of items}
}
\value{
Data frame with flow-factor and sd.
}
\description{
\code{CalculateFlowShortScaleFactors} returns a data frame with flow factors and standard deviation.
}
\examples{
CalculateFlowShortScaleFactors(c(3,4,3,4,2,3,5,6,4,5,6,4,8,9,1,3))
}

