% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalculateFlowShortScaleResults.R
\name{CalculateFlowShortScaleResults}
\alias{CalculateFlowShortScaleResults}
\title{Calculate the flow short scale results.}
\usage{
CalculateFlowShortScaleResults(items, start, end)
}
\arguments{
\item{items}{Questionaire items. A numerical vector of the length 16 of items}

\item{start}{Start time. A scalar of the class 'POSIXct'}

\item{end}{End time. A scalar of the class 'POSIXct'}
}
\value{
List of mean results, sd results and time properties.
}
\description{
\code{CalculateFlowShortScaleResults} returns a list of mean results, sd results and time properties.
}
\examples{
CalculateFlowShortScaleResults(c(3,4,3,4,2,3,5,6,4,5,6,4,8,9,1,3))
CalculateFlowShortScaleResults(c(3,4,3,4,2,3,5,6,4,5,6,4,8,9,1,3))$mean$flow
CalculateFlowShortScaleResults(c(3,4,3,4,2,3,5,6,4,5,6,4,8,9,1,3))$sd$flow

start.time <- as.POSIXct(strptime("2013-10-31 18:11:26", "\%Y-\%m-\%d \%H:\%M:\%S"), tz="CET")
end.time   <- as.POSIXct(strptime("2013-10-31 18:26:26", "\%Y-\%m-\%d \%H:\%M:\%S"), tz="CET")
CalculateFlowShortScaleResults(c(3,4,3,4,2,3,5,6,4,5,6,4,8,9,1,3), start=start.time, end=end.time)$time.properties$time.taken
}

