% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculateInstantaneousPhase.R
\name{CalculateInstantaneousPhase}
\alias{CalculateInstantaneousPhase}
\title{Calculates instantaneous phase of time series in another time series}
\usage{
CalculateInstantaneousPhase(t.1, t.2)
}
\arguments{
\item{t.1.}{First time series in a vector.}

\item{t.2.}{Second time series in a vector.}
}
\value{
A vector with fi.
}
\description{
\code{CalculateInstantaneousPhase} returns a vector with fi
}

