% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculateInstantaneousPhases.R
\name{CalculateInstantaneousPhases}
\alias{CalculateInstantaneousPhases}
\title{Calculates the instantaneous phases of the time series.
\code{CalculateInstantaneousPhases} returns the instantaneous phases of the time series.}
\usage{
CalculateInstantaneousPhases(times, time.series)
}
\arguments{
\item{times}{A numeric vector of the times.}

\item{time.series}{A numeric vector of times of appearance of kth event.}
}
\value{
fi A numeric vector of instantaneous phases.
}
\description{
Calculates the instantaneous phases of the time series.
\code{CalculateInstantaneousPhases} returns the instantaneous phases of the time series.
}

