% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculateJerk.R
\name{CalculateJerk}
\alias{CalculateJerk}
\title{Calculates the jerk of an acceleration.}
\usage{
CalculateJerk(t, x)
}
\arguments{
\item{t.}{Vector with intervals}

\item{x.}{The other vector with acceleration. t and x must have the same length, greater than one, with no missing values.}
}
\value{
A vector with the jerk of the acceleration.
}
\description{
\code{CalculateJerk} returns a vector with the jerk of the acceleration.
}

