% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculateJerkCost.R
\name{CalculateJerkCost}
\alias{CalculateJerkCost}
\title{Calculates the jerk cost of an acceleration in x-, y- and z-direction.}
\usage{
CalculateJerkCost(t, x, y, z, plot = F)
}
\arguments{
\item{t}{Vectors with intervals (s)}

\item{x}{The other vector with acceleration. t and x must have the same length, greater than one, with no missing values.}

\item{y}{The other vector with acceleration. t and y must have the same length, greater than one, with no missing values.}

\item{z}{The other vector with acceleration. t and z must have the same length, greater than one, with no missing values.}

\item{plot}{Boolean. Plot cycle acceleration, cycle jerk and the time differential of jerk-cost}
}
\value{
A vector with the jerk cost [m^2/s^5] of the acceleration.
}
\description{
\code{CalculateJerkCost} returns a vector with the jerk cost of the acceleration.
}

