% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateNormalizedShannonEntropyIndex.R
\name{CalculateNormalizedShannonEntropyIndex}
\alias{CalculateNormalizedShannonEntropyIndex}
\title{Calculates the normalized shannon entropy index of the relative phase of two oscillators}
\usage{
CalculateNormalizedShannonEntropyIndex(t, psi, t.k, t.w = 10, bins = NA)
}
\arguments{
\item{t}{t is a vector with the times of the relative phases.}

\item{psi}{psi is a vector with the realive phase in r (0 to 2pi).}

\item{t.k}{t.k is the point in time for calculting the index.}

\item{t.w}{t.w is the time window around t.k (default = 10).}

\item{bins}{bins is the number of bins (default = The optimal number of bins will be estimated as N = exp[0.626 + 0.4 ln(M - 1)], where M is the number of samples)}
}
\value{
the index.
}
\description{
\code{CalculateNormalizedShannonEntropyIndex} returns the index
}

