% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculateNormalizedShannonEntropyIndexes.R
\name{CalculateNormalizedShannonEntropyIndexes}
\alias{CalculateNormalizedShannonEntropyIndexes}
\title{Calculates the normalized shannon entropy indexes of ?}
\usage{
CalculateNormalizedShannonEntropyIndexes(psi.m, bins = 5)
}
\arguments{
\item{psi.m.}{?.}

\item{bins.}{?.}
}
\value{
A vector of indexes.
}
\description{
\code{CalculateNormalizedShannonEntropyIndexes} returns a vector of indexes
}

