% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculatePhaseAppearances.R
\name{CalculatePhaseAppearances}
\alias{CalculatePhaseAppearances}
\title{Calculates the times of appearance of t of the first time series at the phase of the second time series.
\code{CalculatePhaseAppearances} returns times of appearance of t.}
\usage{
CalculatePhaseAppearances(of, at)
}
\arguments{
\item{of.}{A numeric vector of the first time series.}

\item{at.}{A numeric vector of the second time series.}
}
\value{
fi. A numeric vector of times of appearance of t.
}
\description{
Calculates the times of appearance of t of the first time series at the phase of the second time series.
\code{CalculatePhaseAppearances} returns times of appearance of t.
}

