% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculatePhaseCoherenceIndex.R
\name{CalculatePhaseCoherenceIndex}
\alias{CalculatePhaseCoherenceIndex}
\title{Calculates the phase coherence index of the relative phase of two oscillators}
\usage{
CalculatePhaseCoherenceIndex(t, psi, t.k, t.w = 10)
}
\arguments{
\item{t}{t is a vector with the times of the relative phases.}

\item{psi}{psi is a vector with the realive phase in r (0 to 2pi).}

\item{t.k}{t.k is the point in time for calculting the index.}

\item{t.w}{t.w is the time window around t.k (default = 10).}
}
\value{
the index.
}
\description{
\code{CalculatePhaseCoherenceIndex} returns the index
}

