% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalculatePhaseCoherenceIndexes.R
\name{CalculatePhaseCoherenceIndexes}
\alias{CalculatePhaseCoherenceIndexes}
\title{Calculates the phase coherence indexes from ?}
\usage{
CalculatePhaseCoherenceIndexes(psi, n = 5)
}
\arguments{
\item{psi.}{?.}

\item{n.}{?.}
}
\value{
A vector of indexes.
}
\description{
\code{CalculatePhaseCoherenceIndex} returns a vector of indexes
}

