% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-CalculateRepeatedAnova.R
\name{CalculateRepeatedAnova}
\alias{CalculateRepeatedAnova}
\title{Calculates the repeteated anova of a feature and execute an pairwise t test.}
\usage{
CalculateRepeatedAnova(subject, treatment, feature)
}
\arguments{
\item{subject.}{A vector of subjects}

\item{treatment.}{A factor to group by treatment}

\item{feature.}{A vector of feature values}
}
\description{
\code{CalculateRepeatedAnova} calculates the repeteated anova of a feature and execute an pairwise t test.
}

