% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-CheckFeaturesNormalityInGroups.R
\name{CheckFeaturesNormalityInGroups}
\alias{CheckFeaturesNormalityInGroups}
\title{Checks normality visually and by the Shapiro.}
\usage{
CheckFeaturesNormalityInGroups(features, factor, plot = F)
}
\arguments{
\item{features.}{A vector of features}

\item{factor.}{A factor to group the features}

\item{plot.}{Boolean}
}
\value{
A vector of indexes of normally distributed features
}
\description{
\code{CheckFeaturesNormalityInGroups} checks normality visually and by the Shapiro.
}

