% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{cmd}
\alias{cmd}
\title{Gets vector of command line calls associated with the jobs in the Job object}
\usage{
\S4method{cmd}{Job}(.Object, all = FALSE, quiet = TRUE)
}
\author{
Marcin Imielinski
}

