% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-ComputeFlowShortScaleDimensions.R
\name{ComputeFlowShortScaleDimensions}
\alias{ComputeFlowShortScaleDimensions}
\title{Computes the flow short scale dimensions.}
\usage{
ComputeFlowShortScaleDimensions(items)
}
\arguments{
\item{items.}{A numerical vector of the length 16 of questionaire items}
}
\value{
A data frame with flow dimensions and their standard deviations
}
\description{
\code{ComputeFlowShortScaleDimensions} returns a data frame with flow dimensions and standard deviations.
}
\examples{
ComputeFlowShortScaleDimensions(c(3,4,3,4,2,3,5,6,4,5,6,4,8,9,1,3))
}

