% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-ComputeFScoreTreshold.R
\name{ComputeFScoreTreshold}
\alias{ComputeFScoreTreshold}
\title{Computes a treshold of the f1 score.}
\usage{
ComputeFScoreTreshold(y, p)
}
\arguments{
\item{y.}{A numeric vector with real results}

\item{p.}{A numeric vector with predictions}
}
\value{
The best epsilon "epsilon" and the best f1 score "f1"
}
\description{
\code{ComputeFScoreTreshold} returns best epsilon and best f1 score.
}

