% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-ComputeInstantaneousPhases.R
\name{ComputeInstantaneousPhases}
\alias{ComputeInstantaneousPhases}
\title{Computes the instantaneous phase.}
\usage{
ComputeInstantaneousPhases(ts1, ts2)
}
\arguments{
\item{ts1.}{A numeric vector of time events.}

\item{ts2.}{A numeric vector of time events (onset of the cycle).}
}
\value{
fi A numeric vector of instantaneous phase.
}
\description{
\code{ComputeInstantaneousPhases} returns the instantaneous phase.
}

