% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics-ComputeJerk.R
\name{ComputeJerk}
\alias{ComputeJerk}
\title{Computes the jerk of an acceleration.}
\usage{
ComputeJerk(t, x)
}
\arguments{
\item{t.}{Vector with intervals}

\item{x.}{The other vector with acceleration. t and x must have the same length, greater than one, with no missing values}
}
\value{
A vector with the jerk of the acceleration
}
\description{
\code{ComputeJerk} returns a vector with the jerk of the acceleration.
}

