% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-ComputeMainFrequency.R
\name{ComputeMainFrequency}
\alias{ComputeMainFrequency}
\title{Computes the main frequency of a cyclic signal.}
\usage{
ComputeMainFrequency(y, fs)
}
\arguments{
\item{y.}{A numerical vector of values}

\item{fs.}{A numerical that specifies the sampling rate}
}
\value{
A numerical of main frequency
}
\description{
\code{ComputeMainFrequency} returns the main frequency of a cyclic signal.
}

