% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechnics-ComputeMostCommonCycleMatrix.R
\name{ComputeMostCommonCycleMatrix}
\alias{ComputeMostCommonCycleMatrix}
\title{Computes a matrix of cycles with the most common length.}
\usage{
ComputeMostCommonCycleMatrix(y, separator.indexes)
}
\arguments{
\item{y.}{A numeric vector of data}

\item{separator.indexes.}{A numerical vector of separator indexes}
}
\value{
A matrix of cycles with the most common length
}
\description{
\code{ComputeMostCommonCycleMatrix} return a matrix with cycles the most common length.
}

