% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-ComputeMultivariateGaussian.R
\name{ComputeMultivariateGaussian}
\alias{ComputeMultivariateGaussian}
\title{Computes the probability density function of the multivariate gaussian distribution.}
\usage{
ComputeMultivariateGaussian(X, mu, Sigma2)
}
\arguments{
\item{X.}{A dataset with each n-dimensional data point in one row}

\item{mu.}{The mean mu of the data set}

\item{Sigma2.}{The variances sigma^2, an n x 1 vector}
}
\value{
An n x 1 vector of predictions
}
\description{
\code{ComputeMultivariateGaussian} computes the probability density function of the multivariate gaussian distribution.
}

