% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-ComputeNormalizedShannonEntropyIndex.R
\name{ComputeNormalizedShannonEntropyIndex}
\alias{ComputeNormalizedShannonEntropyIndex}
\title{Computes the normalized shannon entropy index of the relative phase of two oscillators}
\usage{
ComputeNormalizedShannonEntropyIndex(t, psi, t.k, t.w = 10, bins = NA)
}
\arguments{
\item{t}{a numerical vector of the times in second of the relative phases.}

\item{psi}{a numerical vector of the realive phase.}

\item{t.k}{is the point in time for calculting the index.}

\item{t.w}{is the time window around t.k in seconds (default = 10).}

\item{bins}{bins is the number of bins (default = The optimal number of bins will be estimated as N = exp[0.626 + 0.4 ln(M - 1)], where M is the number of samples)}
}
\value{
the index.
}
\description{
\code{ComputeNormalizedShannonEntropyIndex} returns the index.
}

