% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-ComputeOptimalCutoffFrequency.R
\name{ComputeOptimalCutoffFrequency}
\alias{ComputeOptimalCutoffFrequency}
\title{Computes optimal cut off frequency for a low pass butterworth filter}
\usage{
ComputeOptimalCutoffFrequency(x, fs, N, plot = FALSE)
}
\arguments{
\item{x}{a numerical vector.}

\item{fs}{sampling rate.}

\item{N}{filter order.}

\item{plot}{a boolean for control plot.}
}
\value{
the optimal cut off frequency.
}
\description{
\code{ComputeOptimalCutoffFrequency} returns the index.
}

