% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-ComputeRepeatedAnova.R
\name{ComputeRepeatedAnova}
\alias{ComputeRepeatedAnova}
\title{Computes the repeteated anova of a feature and execute an pairwise t test.}
\usage{
ComputeRepeatedAnova(subject, treatment, feature, paired = T)
}
\arguments{
\item{subject.}{A vector of subjects}

\item{treatment.}{A factor to group by treatment}

\item{feature.}{A vector of feature values}

\item{paired.}{A logical indicating whether you want paired t-tests}
}
\description{
\code{ComputeRepeatedAnova} computes the repeteated anova of a feature and execute an pairwise t test.
}

