% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-ComputeSamplingRate.R
\name{ComputeSamplingRate}
\alias{ComputeSamplingRate}
\title{Computes the sampling rate of a time series in milliseconds.}
\usage{
ComputeSamplingRate(t.ms)
}
\arguments{
\item{t.ms.}{A numerical vector of milliseconds}
}
\value{
A numerical with the sampling rate in Hz
}
\description{
\code{ComputeSamplingRate} returns the sampling in Hz.
}

