% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{convert_single_quotes_in_cmd}
\alias{convert_single_quotes_in_cmd}
\title{Cleanup cmd}
\usage{
convert_single_quotes_in_cmd(cmd_original)
}
\description{
Cleaning up cmd.og in Job\@runinfo
}
\details{
For backwards compatibility requires that single quotes in cmd.og are converted to '"'"'
because of bash -c '{ run_cmd ... }' implementation.
}
