% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{cores}
\alias{cores}
\title{Gets cores associated with the jobs in the Job object}
\usage{
\S4method{cores}{Job}(.Object)
}
\description{
Getting LSF / SGE cores associated with Job object
}
\author{
Marcin Imielinski
}
