% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-DetectMidSwings.R
\name{DetectMidSwings}
\alias{DetectMidSwings}
\title{Detects mid swings based on angular velocity.}
\usage{
DetectMidSwings(t.s, angular.velocity.deg.s, fs = 100, cf.1 = 3, cf.2 = 1,
  plot = FALSE)
}
\arguments{
\item{t.s}{a numerical vector of seconds}

\item{angular.velocity.deg.s}{a numerical vector of angular velocity in deg/s.}

\item{fs}{a numerical that specifies the sampling rate (default 100 Hz).}

\item{cf.1}{a numerical that specifies the cut off frequency of the first low pass filter (default = 3 Hz).}

\item{cf.2}{a numerical that specifies the cut off frequency of the second low pass filter (default = 1 Hz).}

\item{plot}{a boolean for control plot.}
}
\value{
A numerical vector of mid swing indexes.
}
\description{
\code{DetectMidSwings} returns mid swings. Uses the angular velocity and filtering like in "Quasi real-time gait event detection using shank-attached gyroscopes" of Lee & Park (2011).
}

