% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{entities}
\alias{entities}
\title{returns original keyed data.table of entities associated with this Job object}
\usage{
\S4method{entities}{Job}(.Object)

\S4method{entities}{Flow}(.Object)
}
\description{
returns original keyed data.table of entities associated with this Job object

returns original keyed data.table of entities associated with this Flow object
}
\author{
Marcin Imielinski
}
