% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{data}
\name{ENV_STRING}
\alias{ENV_STRING}
\title{Default shell command}
\format{
An object of class \code{character} of length 1.
}
\usage{
ENV_STRING
}
\description{
Shell command that executes the script
}
\details{
This is an added command that will enforce the flow command to run from a clean slate.
I.e. no environmental variables inherited from submitted user profile. This ensures,
whether running locally or on slurm, that any environment variables get inherited from user's
bash profile.
}
\keyword{datasets}
