% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-EstimateGaussian.R
\name{EstimateGaussian}
\alias{EstimateGaussian}
\title{Estimates the parameters of a Gaussian distribution using the data in X.}
\usage{
EstimateGaussian(X)
}
\arguments{
\item{X.}{A dataset with each n-dimensional data point in one row}
}
\value{
The mean "mu" of the data set and the variances "sigma2", an n x 1 vector
}
\description{
\code{EstimateGaussian} estimates the parameters of a Gaussian distribution using the data in X.
}

