% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file.name}
\alias{file.name}
\alias{file.dir}
\title{file.name}
\usage{
file.name(paths)

file.dir(paths)
}
\description{
grabs filenames from list of paths

grabs file.dirs from list of paths
}
\author{
Marcin Imielinski

Marcin Imielinski
}
