% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-package.R
\name{flow-options}
\alias{flow-options}
\alias{flow.indenter}
\alias{flow.svg}
\alias{flow.webshot2}
\title{options for the 'flow' package}
\description{
\itemize{
\item \code{flow.indenter}: To circumvent a limitation from nomnoml we use by default
the Braille character \code{"\\u2800"} in places where consecutive standard
spaces would be ignored. On some systems it doesn't print nicely though,
\code{"\\u00b7"} is another option that looks nice.
\item \code{flow.svg}: Whether to use svg graphics in html or viewer output.
\item \code{flow.webshot2}: Whether to use 'webshot2', set to \code{FALSE} to use 'webshot'
as was the default until version \verb{0.2.0}.
}
}
