% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-package.R
\docType{package}
\name{flow-package}
\alias{flow}
\alias{flow-package}
\title{flow: View and Browse Code Using Flow Diagrams}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualize as flow diagrams the logic of functions, expressions or scripts in a static way or when running a call, visualize the dependencies between functions or between modules in a shiny app, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/flow}
  \item \url{https://moodymudskipper.github.io/flow/}
  \item Report bugs at \url{https://github.com/moodymudskipper/flow/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
