% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\docType{package}
\name{flow}
\alias{flow}
\alias{flow-package}
\title{flow: A package for non disruptive flow measures.}
\description{
The flow package provides three categories of important functions:
Questionaire, HRV and Motion.
}
\section{flow functions}{

add
sum
}

