% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_1_flow_run_utils.R
\name{flow_draw}
\alias{flow_draw}
\alias{d}
\title{Draw Diagram From Debugger}
\usage{
flow_draw()

d
}
\value{
Returns \code{NULL} invisibly (called for side effects)
}
\description{
\code{flow_draw()} should only be used in the debugger triggered by a call
to \code{flow_run()}, or following a call to \code{flow_debug()}.
\code{d} is an active binding to \code{flow_draw()}, it means you can just type \code{d}
(without parentheses) instead of \code{flow_draw()}.
}
\details{
\code{d} was designed to look like the other shortcuts detailed in \code{?browser},
such as \code{f}, \code{c} etc... It differs however in that it can be overridden.
For instance if the function uses a variable \code{d} or that a parent environment
contains a variable \code{d}, \code{flow::d} won't be found. In that case you will
have to use \code{flow_draw()}.

If \code{d} or \code{flow_draw()} are called outside of the debugger they will return
\code{NULL} silently.
}
