% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_flow_data.R, R/05_flow_view.R,
%   R/06_flow_run.R, R/07_flow_others.R
\name{flow_data}
\alias{flow_data}
\alias{flow_view}
\alias{flow_run}
\alias{flow_code}
\alias{flow_html}
\alias{flow_browse}
\alias{flow_png}
\title{View function as flow chart}
\usage{
flow_data(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = FALSE)

flow_view(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = FALSE, code = TRUE, width = NULL, height = NULL, ...)

flow_run(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = TRUE, code = TRUE, ..., browse = FALSE)

flow_code(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  code = TRUE, ...)

flow_html(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = FALSE, code = TRUE, width = NULL, height = NULL,
  path = tempfile("flow", fileext = ".html"), ...)

flow_browse(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = FALSE, code = TRUE, width = NULL, height = NULL, ...)

flow_png(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = FALSE, code = TRUE, width = NULL, height = NULL,
  path = tempfile("flow", fileext = ".png"), ...)
}
\arguments{
\item{x}{A call or a function}

\item{range}{The range of boxes to zoom in}

\item{prefix}{Prefix to use for special comments, must start with `"#"`}

\item{sub_fun_id}{if not NULL, the index of the function definition found in
x that we wish to inspect.}

\item{swap}{whether to change `var <- if(cond) expr` into
`if(cond) var <- expr` so the diagram}

\item{code}{Whether to the code in code blocks or only the identifier, to be
more compact, if `NA`, the code will be displayed only if not commented, so
comments can be used as pseudocode for chosen blocks.}

\item{width}{Width in pixels}

\item{height}{height in pixels}

\item{...}{Additional parameters passed to `build_nomnoml_code()`}

\item{browse}{whether to debug step by step (block by block)}

\item{path}{path to save to. By default saves to temp file and prints path.}
}
\description{
`flow_view()` shows the code of a function as a flow diagram, `flow_run()`
runs a call and draws the logical path taken by the code. `flow_png()` and
`flow_html()` are variants of `flow_view()` that allow to print to png or
 html
}
