% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_flow_view_dep.R
\name{flow_view_deps}
\alias{flow_view_deps}
\title{Show dependency tree of a function}
\usage{
flow_view_deps(
  fun,
  max_depth = Inf,
  trim = NULL,
  show_imports = c("functions", "packages", "none"),
  out = NULL
)
}
\arguments{
\item{fun}{A function}

\item{max_depth}{An integer, the maximum depth to display}

\item{trim}{A vector of list of function names where the recursion will stop}

\item{show_imports}{Whether to show imported functions, only packages, or neither.}

\item{out}{a path to save the diagram to.
Special values "html", "htm", "png", "pdf", "jpg" and "jpeg" can be used to
export the objec to a temp file of the relevant format and open it,
if a regular path is used the format will be guessed from the extension.}
}
\description{
Show dependency tree of a function
}
