% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_flow_view_uses.R
\name{flow_view_uses}
\alias{flow_view_uses}
\title{Show graph of callers of a function}
\usage{
flow_view_uses(x, pkg = NULL, out = NULL)
}
\arguments{
\item{x}{An object}

\item{pkg}{A package or environment to fetch callers from, by default \code{fun}'s environment}

\item{out}{a path to save the diagram to.
Special values "html", "htm", "png", "pdf", "jpg" and "jpeg" can be used to
export the object to a temp file of the relevant format and open it,
if a regular path is used the format will be guessed from the extension.}
}
\value{
\code{flow_view_uses()} returns a \code{"flow_diagram"} object by default, and the output path invisibly if \code{out} is not
\code{NULL} (called for side effects).
}
\description{
Experimental function that displays for a given object or function all functions
that call it directly or indirectly.
}
\details{
The function is not very robust yet, but already useful for many usecases.
}
\examples{
flow_view_uses(flow_run)
}
