% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_flow_view_vars.R
\name{flow_view_vars}
\alias{flow_view_vars}
\title{Draw the dependencies of variables in a functiomn}
\usage{
flow_view_vars(fun)
}
\arguments{
\item{fun}{The function to draw}
}
\value{
Called for side effects
}
\description{
Experimental, and probably can never been made 100\% robust
}
