% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{FlowAnnotation-class}
\alias{FlowAnnotation-class}
\title{Class to represent an entity-specific annotation (ie a path or value), used for wiring modules into tasks}
\description{
Class \code{FlowAnnotation} is a simple container for storing arguments to Task that are to be interpreted as annotations
used when initializing a task config.
S4 class for \code{FlowAnnotation}
}
\section{Slots}{

\describe{
  \item{arg}{scalar character argument that will be interpreted literally}
}
}

