% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\name{FlowAnnotation}
\alias{FlowAnnotation}
\title{Creates an entity-specific annotation object (ie a path or value), used for wiring modules into tasks}
\usage{
FlowAnnotation(...)
}
\arguments{
\item{name}{character name of argument}

\item{arg}{value of argument}

\item{path}{boolean specifying whether this is a path or not}

\item{default}{can specify default value (ie what do to if annotation is missing from input table) (default is NULL)}
}
\description{
Creates an entity-specific annotation object (ie a path or value), used for wiring modules into tasks
}

