% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{FlowLiteral-class}
\alias{FlowLiteral-class}
\title{Class to represent a literal argument to a task (ie a static path or value), used for wiring modules into tasks}
\description{
S4 class for \code{FlowLiteral}

Class \code{FlowLiteral} is a simple container for storing arguments to Task that are to be interpreted literally
used when initializing a task config.
}
\section{Slots}{

\describe{
  \item{arg}{scalar character argument that will be interpreted literally}
}
}
\author{
Marcin Imielinski
}

