% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{FlowLiteral}
\alias{FlowLiteral}
\title{Creates object representing a literal argument to a task (ie a static path or value), used for wiring modules into tasks}
\usage{
FlowLiteral(...)
}
\arguments{
\item{name}{character name of argument}

\item{arg}{value of argument}

\item{path}{boolean specifyign whether this is a path or not}
}
\author{
Marcin Imielinski
}

