% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{FlowOutput-class}
\alias{FlowOutput-class}
\title{Class representing the wiring of a regexp on a job output directory to an output annotation
S4 class for \code{FlowOutput}

Class \code{FlowOutput}is a simple container for capturing  output from Tasks
which consists of name (of output annotation) and pattern (to match the output path in the
task's output directory)}
\section{Slots}{

\describe{
  \item{name}{scalar character specifying annotation to write}
  \item{pattern}{scalar character regexp pattern to match the file in the output directory}
}
}

