% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{FlowOutput}
\alias{FlowOutput}
\title{Constructs an object representing the wiring of a regexp on a job output directory to an output annotation}
\usage{
FlowOutput(...)
}
\arguments{
\item{name}{annotation name to attach output file}

\item{pattern}{pattern to match to get output file in output directory}
}
\description{
Constructs an object representing the wiring of a regexp on a job output directory to an output annotation
}
\author{
Marcin Imielinski
}
