% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-GenericWrite.R
\name{GenericWrite}
\alias{GenericWrite}
\title{Writes data with a write function in a file .}
\usage{
GenericWrite(x, file, header, footer, f = write.csv, ...)
}
\arguments{
\item{...}{write options.}

\item{x.}{The data to write}

\item{file.}{The File to write in}

\item{header.}{A extra header on top of the data}

\item{footer.}{A extra footer below the data}

\item{f.}{A write function}
}
\description{
\code{GenericWrite} writes data with a write function in a file.
}

