% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getslotchain}
\alias{getslotchain}
\title{Get chain of S4 methods}
\usage{
getslotchain(object, ..., default = NULL)
}
\description{
Function that goes through the chain of arguments and returns a default
instead of erroring out
}
