% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{gres}
\alias{gres}
\title{Gets queue or partition associated with the jobs in the Job object}
\usage{
\S4method{gres}{Job}(.Object)
}
\description{
Getting SLURM  gres associated with Job object
}
\author{
Marcin Imielinski
}
