% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{ids}
\alias{ids}
\title{returns vector of entity ids associated with this Job objedct}
\usage{
\S4method{ids}{Job}(.Object)
}
\description{
returns vector of entity ids associated with this Job objedct
}
\author{
Marcin Imielinskix
}

