% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{inputs}
\alias{inputs}
\title{returns keyed data.table of inputs annotation associated with the jobs in this Job object}
\usage{
\S4method{inputs}{Job}(.Object)
}
\description{
returns keyed data.table of inputs annotation associated with the jobs in this Job object
}
\author{
Marcin Imielinski
}

