% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{io_n<-}
\alias{io_n<-}
\title{Sets ionice parameters associated with Job object}
\usage{
\S4method{io_n}{Job}(.Object) <- value
}
\description{
setting io_n parameters associated with job object
}
\author{
Kevin Hadi
}
