% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{jname}
\alias{jname}
\title{Returns character vector of job names associated with this Job object.}
\usage{
\S4method{jname}{Job}(.Object)
}
\description{
Returns character vector of job names associated with this Job object.
}
\author{
Marcin Imielinski
}

