% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flowfield.R
\name{make_flowfield}
\alias{make_flowfield}
\title{Create a flowfied}
\usage{
make_flowfield(
  width = 100,
  height = 100,
  noise = c("perlin", "simplex", "cubic", "value", "waves", "spheres", "white",
    "checkerboard"),
  angle = 1,
  octaves = 1,
  frequency = 0.01,
  fractal = c("fbm", "billow", "clamped", "ridged"),
  seed = runif(1, max = 10000),
  plot = FALSE
)
}
\arguments{
\item{plot}{}
}
\value{
A data.frame with three columns: x and y coordinates, and angle, the
angle of flow at that grid coordinate
}
\description{
Create a flowfied
}
