% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenly_spaced_trails.R
\name{make_trail_x}
\alias{make_trail_x}
\title{Title}
\usage{
make_trail_x(flowfield_df, max_steps = 100, dsep = 0.05, dtest = 0.5)
}
\arguments{
\item{flowfield_df}{A data.frame containing rows with x and y coordinates and
a corresponding angle.}

\item{max_steps}{The maximum number of steps (i.e. increments of \code{dsep})
that a line is allowed to grow. (Of course, a line will end earlier if it
hits the boundary of the flowfield or gets too close to another line
first.)}

\item{dsep}{The separating distance of newly-generated lines, as a proportion
of the width of the flowfield.}

\item{dtest}{The minimal distance that lines are allowed to converge. A
proportion of \code{dsep}.}
}
\value{

}
\description{
Title
}
