% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trails.R
\name{make_trails}
\alias{make_trails}
\title{Generate flowfield trails}
\usage{
make_trails(
  flowfield_df,
  particles = 100,
  start_positions = c("grid", "runif", "poisson"),
  max_steps = 1,
  step_length = 0.05,
  direction = c("both", "forward", "backwards"),
  dtest = 0
)
}
\arguments{
\item{dtest}{}
}
\value{

}
\description{
Generate flowfield trails
}
