% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{mem<-}
\alias{mem<-}
\title{Sets max memory associated with the jobs in the Job object}
\usage{
\S4method{mem}{Job}(.Object) <- value
}
\description{
Setting max mem (in GB) associated with Job object
}
\author{
Marcin Imielinski
}
