% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{module-task}
\alias{module-task}
\title{Gets module associated with a Task}
\usage{
\S4method{module}{Task}(.Object)
}
\description{
getting module associated with Task object
}
\author{
Marcin Imielinski
}

