% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{more}
\alias{more}
\title{more}
\usage{
more(x, grep = NULL, pipe = FALSE)
}
\arguments{
\item{x}{vector of iles}

\item{grep}{string to grep in files (=NULL)}
}
\description{
"more" +/- grep vector of files
}
\author{
Marcin Imielinski
}
