% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-MultivariateGaussian.R
\name{MultivariateGaussian}
\alias{MultivariateGaussian}
\title{MultivariateGaussian computes the probability density function of the multivariate gaussian distribution.}
\usage{
MultivariateGaussian(X, mu, Sigma2)
}
\description{
\code{MultivariateGaussian} computes the probability density function of the multivariate gaussian distribution.
density function of the examples X under the multivariate gaussian distribution with parameters mu and Sigma2. If Sigma2 is a matrix, it is treated as the covariance matrix. If Sigma2 is a vector, it is treated as the \sigma^2 values of the variances in each dimension (a diagonal covariance matrix)
}

