% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{normalizePath}
\alias{normalizePath}
\title{normalizePath}
\usage{
normalizePath(x)
}
\description{
Internal version of normalizePath that doesn't dereference symbolic links to the file proper
instead just to the containing directory
}
