% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{now<-}
\alias{now<-}
\title{Sets now param associated with the jobs in the Job object}
\usage{
\S4method{now}{Job}(.Object) <- value
}
\description{
Setting LSF / SGE now associated with Job object
}
\author{
Marcin Imielinski
}
