% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{outputs}
\alias{outputs}
\title{returns keyed data.table of output entity annotations generated from the jobs in this Job object.  Annotations will be NA for non-completed jobs.}
\usage{
\S4method{outputs}{Job}(.Object)
}
\author{
Marcin Imielinski
}

