% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse.info}
\alias{parse.info}
\title{parse.info}
\usage{
parse.info(jname, detailed = F, force = FALSE, mc.cores = 1)
}
\description{
parses outputs of completed Job by subdirectory
}
\author{
Marcin Imielinski
}
