% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-PlotFeaturesInGroups.R
\name{PlotFeaturesInGroups}
\alias{PlotFeaturesInGroups}
\title{Plots groups of features as histogram and boxplat.}
\usage{
PlotFeaturesInGroups(features, factor)
}
\arguments{
\item{features.}{A vector of features}

\item{factor.}{A factor to group the features}
}
\description{
\code{PlotFeaturesInGroups} plots groups of features as histogram and boxplat.
}

