% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-PlotMostCommonCycleAcceleration.R
\name{PlotMostCommonCycleAcceleration}
\alias{PlotMostCommonCycleAcceleration}
\title{Plots cycles of acceleration with the most common length.}
\usage{
PlotMostCommonCycleAcceleration(a, separator.indexes, ...)
}
\arguments{
\item{a.}{A numeric vector of acceleration}

\item{separator.indexes.}{A numerical vector of separator indexes}
}
\value{
A numeric vector of the mean cycle with the most common length
}
\description{
\code{PlotMostCommonCycleAcceleration} plots cycles of acceleration with the most common length.
}

