% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{qcmd}
\alias{qcmd}
\title{Returns vector of qsub (ie SGE) commands associated with this Job object}
\usage{
\S4method{qcmd}{Job}(.Object, all = FALSE)
}
\arguments{
\item{all}{logical flag whether to run all jobs (including completed)}
}
\description{
Returns vector of qsub (ie SGE) commands associated with this Job object
}
\author{
Marcin Imielinski
}
