% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{qjobs}
\alias{qjobs}
\title{Tracks down any sGE jobs associated with this Job object using a bjobs query (warning can be slow)}
\usage{
\S4method{qjobs}{Job}(.Object)
}
\description{
Tracks down any sGE jobs associated with this Job object using a bjobs query (warning can be slow)
}
\author{
Marcin Imielinski
}
