% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{qos<-}
\alias{qos<-}
\title{Sets qos associated with the jobs in the Job object}
\usage{
\S4method{qos}{Job}(.Object) <- value
}
\description{
Setting LSF / SGE qos associated with Job object
}
\author{
Marcin Imielinski
}
