% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{qprior<-}
\alias{qprior<-}
\title{Sets qprior parameter associated with job object}
\usage{
\S4method{qprior}{Job}(.Object) <- value
}
\description{
Sets qprior parameter which passes a value onto qsub -p command
}
\author{
Kevin Hadi
}
