% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{qrun}
\alias{qrun}
\title{Runs jobs on LSF}
\usage{
\S4method{qrun}{Job}(.Object, mc.cores = 1, all = FALSE)
}
\arguments{
\item{mc.cores}{Number of parallel cores to run jobs with (=1)}
}
\description{
Runs jobs on LSF
}
\author{
Marcin Imielinski
}
