% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qsub_cmd}
\alias{qsub_cmd}
\title{qsub_cmd}
\usage{
qsub_cmd(script.fn, queue, jname = NULL, jlabel = NULL, jgroup = NULL,
  mem = NULL, group = "cgafolk", cwd = NULL, mc.cores = NULL,
  deadline = F, now = FALSE)
}
\description{
Makes qsub command that wraps shell command "cmd" to send to queue "queue"
redirecting output / error etc streams to path prefixed by "jname",
optional_args: maximum memory requirements "mem", "jlabel" job label
}
\author{
Marcin Imielinski
}
