% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{queue<-}
\alias{queue<-}
\title{Sets queue associated with the jobs in the Job object}
\usage{
\S4method{queue}{Job}(.Object) <- value
}
\description{
Setting LSF / SGE queue associated with Job object
}
\author{
Marcin Imielinski
}
