% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_insert_redraw.R
\name{redraw}
\alias{redraw}
\title{redraw}
\usage{
redraw(always = FALSE)
}
\arguments{
\item{always}{not implemented yet, set to TRUE to always redraw automatically during current run}
}
\description{
redraw
}
