% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{refresh}
\alias{refresh}
\title{refreshes .rds copy of job object from standard location (TASK.NAME.rds in Job output root directory)}
\usage{
\S4method{refresh}{Job}(object, verbose = TRUE)
}
\author{
Marcin Imielinski
}
