% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{report}
\alias{report}
\title{Retrieves a data.table reporting detailed runtime stats associated with each job in the Job object.}
\usage{
\S4method{report}{Job}(.Object, mc.cores = 1, force = FALSE)
}
\description{
Scrapes the underlying output directories associated with this Job object to determine whether jobs completed, stats on how much
memory and time was used by the job, when the jobs launched and when they finished, etc.

This is used by the Job update method, but can also be useful for debugging and diagnostics.
}
\author{
Marcin Imielinski
}

