% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset.job}
\alias{reset.job}
\title{reset.job

Reset a job with different params}
\usage{
reset.job(
  x,
  delete_cache = FALSE,
  ...,
  i = NULL,
  rootdir,
  jb.mem,
  jb.cores,
  jb.time,
  update_cores = 1,
  shell,
  force_shell,
  force_profile,
  task = NULL
)
}
\value{
A Flow job object
}
\description{
reset.job

Reset a job with different params
}
\author{
Kevin Hadi
}
