% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-ReturnFieldLabels.R
\name{ReturnFieldLabels}
\alias{ReturnFieldLabels}
\title{Returns label for the fields.}
\usage{
ReturnFieldLabels(col.name)
}
\arguments{
\item{col.name.}{A character vector that specifies the column name}
}
\value{
A expression for the field label
}
\description{
\code{ReturnFieldLabels} returns the field label by column
}

