% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{run}
\alias{run}
\title{Runs jobs as local commands from shell via R system calls}
\usage{
\S4method{run}{Job}(.Object, mc.cores = 1, all = FALSE, quiet = TRUE)
}
\arguments{
\item{mc.cores}{Number of parallel cores to run jobs with (=1)}
}
\author{
Marcin Imielinski
}

