% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{scmd}
\alias{scmd}
\title{Returns vector of ssub (ie Slurm) commands associated with this Job object}
\usage{
\S4method{scmd}{Job}(.Object, all = FALSE)
}
\arguments{
\item{all}{logical flag whether to run all jobs (including completed)}
}
\description{
Returns vector of ssub (ie Slurm) commands associated with this Job object
}
\author{
Zoran Gajic
}
