% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-SearchExtrema.R
\name{SearchExtrema}
\alias{SearchExtrema}
\title{Search for extrema in a vector.}
\usage{
SearchExtrema(v, which = c("both", "maxima", "minima"))
}
\arguments{
\item{v}{Vector of a function}

\item{which}{One of "both", "maxima", "minima". Default is "both"}
}
\value{
Vector of extrema indices
}
\description{
\code{SearchExtrema} Search for extrema in a vector.
}

