% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silent}
\alias{silent}
\title{run expression without any printed output

execute expression without any output to console.
silent({var = function_that_has_explicit_print(...)})}
\usage{
silent(this_expr, this_env = parent.frame())
}
\arguments{
\item{...}{an expression}
}
\description{
run expression without any printed output

execute expression without any output to console.
silent({var = function_that_has_explicit_print(...)})
}
\author{
Kevin Hadi
}
