% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{skill,Job-method}
\alias{skill,Job-method}
\title{#' @name skill
#' @title Kills any running Slurm jobs associated with this Job object
#' @exportMethod skill
#' @export
#' @author Andre Forbes}
\usage{
\S4method{skill}{Job}(.Object, jid = NULL)
}
\description{
#' @name skill
#' @title Kills any running Slurm jobs associated with this Job object
#' @exportMethod skill
#' @export
#' @author Andre Forbes
}
