% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{Slurmize}
\alias{Slurmize}
\title{updates cmd.sh to run on slurm}
\usage{
Slurmize(job)
}
\arguments{
\item{job, }{the job object to update}
}
\description{
updates cmd.sh to run on slurm
}
\author{
Zoran Gajic
}
