% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{status.info}
\alias{status.info}
\title{returns vector of more detailed statuses associated with jobs in this Job object}
\usage{
\S4method{status.info}{Job}(.Object)
}
\description{
returns vector of more detailed statuses associated with jobs in this Job object
}
\author{
Marcin Imielinski
}
