% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{status}
\alias{status}
\title{Returns character vector of status of jobs in this Job object.  Regexps on these statuses can be also used as character inputs to [] accessor}
\usage{
\S4method{status}{Job}(.Object)
}
\author{
Marcin Imielinski
}

