% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{[.Flow}
\alias{[.Flow}
\title{Flow subset}
\usage{
\method{[}{Flow}(obj, i = NULL, j = NULL, include = NULL, exclude = NULL, quick = FALSE, ...)
}
\arguments{
\item{obj}{Flow object to be subset}

\item{i}{indices of entities}

\item{j}{indices of tasks to subset}

\item{include}{character vector representing tasks to include in output object}

\item{exclude}{character vector representing nodes or tasks to exclude in output object}
}
\value{
A new gNode that contains only the given id's
}
\description{
Overloads subset operator for Flow object. Allow subsetting object based on entities
features (first argument, a la "rows") and tasks (a la "columns").

First argument can be an expression on entity metadata that returns an integer or logical
of length = nrow(entities).

Tasks can be indices or names of tasks
}
\author{
Marcin Imielinski
}
