% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tailf}
\alias{tailf}
\title{tailf}
\usage{
tailf(x, n = NULL, grep = NULL)
}
\arguments{
\item{x}{vector of iles}

\item{grep}{string to grep in files (=NULL)}
}
\description{
"tail -f" +/- grep vector of files
}
\author{
Marcin Imielinski
}
