% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{time<-}
\alias{time<-}
\title{Sets max time limit associated with the jobs in the Job object}
\usage{
\S4method{time}{Job}(.Object) <- value
}
\description{
Setting time limit ("days-hours" or "hours") associated with Job object
e.g. ("3-00" or "72" for 3 days / 72 hours)
}
\author{
Kevin Hadi
}
