% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{time}
\alias{time}
\title{Gets time associated with the jobs in the Job object}
\usage{
\S4method{time}{Job}(.Object)
}
\description{
Getting time associated with Job object
}
\author{
Zoran Gajic
}
