% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{update}
\alias{update}
\title{updates Job status

This method checks inputs for updates to input files, polls output directories for the presence output files, checks stdout output logs, among
other things to assess job status.  Once finished, it caches results to the .rds object associated with this Job.}
\usage{
\S4method{update}{Job}(object, check.inputs = TRUE, mc.cores = 1,
  cache.object = TRUE, print.status = TRUE)
}
\description{
updates Job status

This method checks inputs for updates to input files, polls output directories for the presence output files, checks stdout output logs, among
other things to assess job status.  Once finished, it caches results to the .rds object associated with this Job.
}
\author{
Marcin Imielinski
}
