% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_flow.R
\name{view_flow}
\alias{view_flow}
\title{View function as flow chart}
\usage{
view_flow(
  f,
  prefix = NULL,
  code = TRUE,
  png = NULL,
  width = NULL,
  height = NULL,
  output = c("view", "nomnoml", "data"),
  ...,
  svg = FALSE
)
}
\arguments{
\item{f}{}

\item{prefix}{prefix to use for special comments, must start with `"#"`}

\item{...}{additional parameters passed to `build_nomnoml_code()`}
}
\description{
View function as flow chart
}
