% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WriteDataSubsetInCSVFile.R
\name{WriteDataSubsetInCSVFile}
\alias{WriteDataSubsetInCSVFile}
\title{Write subsets of data in csv files with incremented index.}
\usage{
WriteDataSubsetInCSVFile(data, times, starts, ends, file.name, file.path)
}
\arguments{
\item{data}{Data. A data frame}

\item{times}{Times. A vector of the class 'POSIXct' with the same length row number than data}

\item{starts}{Start time of each subset. A vector of the class 'POSIXct' with start times  (must have the same length than 'ends')}

\item{ends}{End time of each subset. A vector of the class 'POSIXct' with end times (must have the same length than 'starts')}

\item{file.name}{File name. A vector of the class 'Character' with the file name}

\item{path.name}{Path name. A vector of the class 'Character' with the path name, where the csv files should be written}
}
\description{
\code{WriteDataSubsetInCSVFile} Writes subsets of data in csv files with incremented index.
}

