% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{xml2task}
\alias{xml2task}
\title{Makes a best attempt to convert a firehose xml task configuration into a Task object or .task file}
\usage{
xml2task(path, module = NULL, out.file = NULL)
}
\description{
Takes a path to an xml file of a FH task configuration, module path, and outputs a Task object or writes
to a .task fil.e
}
\author{
Marcin Imielinski
}

