% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{bjobs}
\alias{bjobs}
\title{Tracks down any LSF jobs associated with this Job object using a bjobs query (warning can be slow)}
\usage{
\S4method{bjobs}{Job}(.Object, mc.cores = 1)
}
\description{
Tracks down any LSF jobs associated with this Job object using a bjobs query (warning can be slow)
}
\author{
Marcin Imielinski
}

