% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-CheckMidSwings.R
\name{CheckMidSwings}
\alias{CheckMidSwings}
\title{CheckMidSwings}
\usage{
CheckMidSwings(t.ms, angular.velocity.deg.s, length.s, mid.swing.indexes)
}
\description{
\code{CheckMidSwings} is ...
}

