% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-CheckMidSwings.R
\name{CheckMidSwings}
\alias{CheckMidSwings}
\title{Checks mid swing visually.}
\usage{
CheckMidSwings(t.s, angular.velocity.deg.s, mid.swing.indexes)
}
\arguments{
\item{t.s.}{t.s. A numerical vector of seconds}

\item{angular.velocity.deg.s.}{A numerical vector of angular velocity in deg/s}

\item{mid.swing.indexes.}{A numerical vector of mid swing indexes}
}
\value{
A numerical vector of checked mid swing indexes
}
\description{
\code{CheckMidSwings} return checked mid swing indexes.
}

