% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-ComputeOptimalCutoffFrequency.R
\name{ComputeOptimalCutoffFrequency}
\alias{ComputeOptimalCutoffFrequency}
\title{Computes optimal cut off frequency for a low pass butterworth filter}
\usage{
ComputeOptimalCutoffFrequency(x, fs, N)
}
\arguments{
\item{x}{vector with data.}

\item{fs}{sampling rate.}

\item{N}{Order of the filter.}
}
\value{
the optimal cut off frequency.
}
\description{
\code{ComputeOptimalCutoffFrequency} returns the index.
}

