% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-ComputePhaseCoherenceIndex.R
\name{ComputePhaseCoherenceIndex}
\alias{ComputePhaseCoherenceIndex}
\title{Computes the phase coherence index of the relative phase of two oscillators}
\usage{
ComputePhaseCoherenceIndex(t, psi, t.k, t.w = 10)
}
\arguments{
\item{t}{t is a vector with the times of the relative phases.}

\item{psi}{psi is a vector with the realive phase.}

\item{t.k}{t.k is the point in time for calculting the index.}

\item{t.w}{t.w is the time window around t.k (default = 10).}
}
\value{
the index.
}
\description{
\code{ComputePhaseCoherenceIndex} returns the index.
}

