% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-DetectAnomaly.R
\name{DetectAnomaly}
\alias{DetectAnomaly}
\title{Detects anomaly of two variables.}
\usage{
DetectAnomaly(x, y, x.lab, y.lab, x.lim, y.lim, epsilon = 0, ...)
}
\arguments{
\item{x}{a numeric vector of the first variable.}

\item{y}{a numeric vector of the second variable.}

\item{epsilon}{a numeric that specifies a new epsilon (default = 0).}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Epsilon "epsilon" and a numeric vector with indexes of outliers "outliers".
}
\description{
\code{DetectAnomaly} returns outliers and a new epsilon for a new iteration.
}

