% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomechanics-DetectMidSwings.R
\name{DetectMidSwings}
\alias{DetectMidSwings}
\title{DetectMidSwings}
\usage{
DetectMidSwings(t.s, angular.velocity, fs = 102.4, ff.1 = 4, ff.2 = 0.5)
}
\description{
\code{DetectMidSwings} is ...
}

