% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{dirs}
\alias{dirs}
\title{returns list of character vectors containing the contents of directory associated with these jobs}
\usage{
\S4method{dirs}{Job}(.Object, full = TRUE, ...)
}
\description{
returns list of character vectors containing the contents of directory associated with these jobs
}
\author{
Marcin Imielinski
}

