% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{.update_cmd}
\alias{.update_cmd}
\title{Create Flow command}
\usage{
.update_cmd(.Object, qos = NULL, ...)
}
\description{
Instantiates flow command
}
\details{
Flow command is instantiated with parameters
 and accessed via Flow::cmd().
}
