% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{entities}
\alias{entities}
\title{returns original keyed data.table of entities associated with this Job object}
\usage{
\S4method{entities}{Job}(.Object)
}
\description{
returns original keyed data.table of entities associated with this Job object
}
\author{
Marcin Imielinski
}
