% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-package.R
\docType{package}
\name{flow-package}
\alias{flow}
\alias{flow-package}
\title{flow: View and Browse Code Using Flow Diagrams}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Visualize as flow diagrams the logic of functions, expressions or scripts in a static way or when running a call, and ease debugging. Advanced features include analogs to 'debug' and 'debugonce' to target specific functions to draw, an utility to draw the calls used in the tests of the package in a markdown report, and an utility to draw all the functions of one package in a markdown report.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/flow}
  \item \url{https://moodymudskipper.github.io/flow/}
  \item Report bugs at \url{https://github.com/moodymudskipper/flow/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
