% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-package.R
\docType{package}
\name{flow-package}
\alias{flow}
\alias{flow-package}
\title{flow: Create Flowfields}
\description{
Create a flowfield, populate it with trails, and visualize the result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robbrotherton/flow}
  \item Report bugs at \url{https://github.com/robbrotherton/flow/issues}
}

}
\author{
\strong{Maintainer}: Rob Brotherton \email{robert.brotherton@gmail.com}

}
\keyword{internal}
