% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{flow}
\alias{flow}
\title{flow}
\description{
Create a flowfield, populate it with trails, and visualize the result
}
\author{
Rob Brotherton \href{mailto:robert.brotherton@gmail.com}{robert.brotherton@gmail.com}
}
