% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_flow_doc.R
\name{flow_doc}
\alias{flow_doc}
\title{Draw Flow Diagrams for an Entire Package}
\usage{
flow_doc(pkg, out = paste0(pkg, ".html"), ...)
}
\arguments{
\item{pkg}{package name as a string}

\item{out}{path to an html file to write to}

\item{...}{additional arguments passed to flow_view}
}
\description{
Draw Flow Diagrams for an Entire Package
}
