% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_flow_data.R, R/05_flow_view.R,
%   R/06_flow_run.R
\name{flow_data}
\alias{flow_data}
\alias{flow_view}
\alias{flow_run}
\title{View function as flow chart}
\usage{
flow_data(
  x,
  range = NULL,
  prefix = NULL,
  sub_fun_id = NULL,
  swap = TRUE,
  narrow = FALSE
)

flow_view(
  x,
  range = NULL,
  prefix = NULL,
  sub_fun_id = NULL,
  swap = TRUE,
  narrow = FALSE,
  code = TRUE,
  width = NULL,
  height = NULL,
  ...,
  out = NULL,
  svg = FALSE,
  engine = c("nomnoml", "plantuml")
)

flow_run(
  x,
  prefix = NULL,
  swap = TRUE,
  code = TRUE,
  ...,
  out = NULL,
  svg = FALSE,
  browse = FALSE,
  show_passes = FALSE
)
}
\arguments{
\item{x}{A call, a function, or a path to a script}

\item{range}{numeric vector used to compute the range of boxes to zoom in}

\item{prefix}{prefix to use for special comments in our code used as block headers,
must start with \code{"#"}}

\item{sub_fun_id}{if not NULL, the index or name of the function definition found in
x that we wish to inspect}

\item{swap}{whether to change \code{var <- if(cond) expr} into
\code{if(cond) var <- expr} so the diagram displays better}

\item{narrow}{\code{TRUE} makes sure the diagram stays centered on one column
(they'll be longer but won't shift to the right)}

\item{code}{Whether to display the code in code blocks or only the header, to be
more compact, if \code{NA}, the code will be displayed only if no header is
defined by special comments.}

\item{width, height}{Width and height in pixels, passed to \code{htmlwidgets::createWidget()}}

\item{...}{Additional parameters passed to \code{build_nomnoml_code()}}

\item{out}{a path to save the diagram to.
Special values "html", "htm", "png", "pdf", "jpg" and "jpeg" can be used to
export the objec to a temp file of the relevant format and open it,
if a regular path is used the format will be guessed from the extension.}

\item{svg}{only for default or html outut, whether to use svg rendering, rendering
is more robust without it, but it makes text selectable on output which is
sometimes useful}

\item{engine}{Either \code{"nomnoml"} (default) or \code{"plantuml"} (experimental), if
the latter, arguments \code{range}, \code{prefix}, \code{narrow}, \code{code}, \code{width},
\code{height} and \code{...} will be ignored.}

\item{browse}{whether to debug step by step (block by block),
can also be a vector of block ids, in this case \code{browser()} calls will be
inserted at the start of these blocks}

\item{show_passes}{label the edges with the number of passes}
}
\description{
\code{flow_view()} shows the code of a function as a flow diagram, \code{flow_run()}
runs a call and draws the logical path taken by the code, \code{flow_data()} is
the lower level function, which builds the edge and node data rendered by
\code{flow_view()} and \code{flow_run()}.
}
