% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_flow_data.R, R/05_flow_view.R,
%   R/06_flow_run.R
\name{flow_data}
\alias{flow_data}
\alias{flow_view}
\alias{flow_run}
\title{View function as flow chart}
\usage{
flow_data(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = TRUE, narrow = FALSE)

flow_view(x, range = NULL, prefix = NULL, sub_fun_id = NULL,
  swap = TRUE, narrow = FALSE, code = TRUE, width = NULL,
  height = NULL, ..., out = NULL, svg = FALSE)

flow_run(x, prefix = NULL, swap = TRUE, code = TRUE, ...,
  out = NULL, svg = FALSE, browse = FALSE, trim = FALSE)
}
\arguments{
\item{x}{A call or a function}

\item{range}{The range of boxes to zoom in}

\item{prefix}{Prefix to use for special comments, must start with `"#"`}

\item{sub_fun_id}{if not NULL, the index of the function definition found in
x that we wish to inspect.}

\item{swap}{whether to change `var <- if(cond) expr` into
`if(cond) var <- expr` so the diagram}

\item{narrow}{make sure diagrams stay centered on one column (they'll be
longer but won't shift to the right)}

\item{code}{Whether to the code in code blocks or only the identifier, to be
more compact, if `NA`, the code will be displayed only if not commented, so
comments can be used as pseudocode for chosen blocks.}

\item{width}{Width in pixels}

\item{height}{height in pixels}

\item{...}{Additional parameters passed to `build_nomnoml_code()`}

\item{out}{a path to save the diagram to.
Special values "html", "htm", "png", "pdf", "jpg" and "jpeg" can be used to
export the objec to a temp file of the relevant format and open it,
if a regular path is used the format will be guessed from the extension.}

\item{svg}{only for default or html outut, whether to use svg rendering, rendering
is more robust without it, but it makes text selectable on output which is
 sometimes useful.}

\item{browse}{whether to debug step by step (block by block)}

\item{trim}{wether to trim the head of the diagram, only relevant if browse
is `TRUE`}
}
\description{
`flow_view()` shows the code of a function as a flow diagram, `flow_run()`
runs a call and draws the logical path taken by the code, `flow_data()` is
the lower level function, which builds the edge and node data rendered by
`flow_view()` and `flow_run()`.
}
