% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{key}
\alias{key}
\title{Retrieves the entity key associated with this Job object}
\usage{
\S4method{key}{Job}(.Object)
}
\description{
Retrieves the entity key associated with this Job object
}
\author{
Marcin Imielinski
}
