% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trails.R
\name{make_trails}
\alias{make_trails}
\title{Title}
\usage{
make_trails(
  flowfield_df,
  particles = 100,
  start_positions = c("grid", "runif", "poisson"),
  steps = 0.1,
  step_length = 0.05,
  min_dist = 0
)
}
\arguments{
\item{particles}{}
}
\value{

}
\description{
Title
}
