% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trails.R
\name{make_trails}
\alias{make_trails}
\title{Generate flowfield trails}
\usage{
make_trails(
  flowfield,
  particles = trail_seeds(100, ff_limits(flowfield)),
  step_length = 1,
  direction = c("both", "forward", "backward"),
  existing_trails_df = NULL
)
}
\arguments{
\item{flowfield}{A data frame, or a list of data frames. Each flowfield data
frame must contain three columns, x, y, and a: the coordinates of the
flowfield grid and the angle associated with each grid reference. If a list
of data frames is supplied, trails will be drawn in alternating order.}

\item{particles}{A data frame containing two columns, x and y: the
coordinates of the starting points (particles) from which trails will be
generated on the flowfield.}

\item{step_length}{The length of each 'step' taken as a trail is generated.
Smaller values result in smoother lines.}

\item{direction}{Which direction should a trail be drawn from a seedpoint?
One of 'forward', 'backward' or 'both'.}

\item{existing_trails_df}{A data frame containing previously-drawn trails.
New trails will avoid collisions with these according to dtest.}

\item{max_steps}{The maximum number of steps to take from each trail seed
point. Trails will be terminated prior to taking this many steps if the go
out of bounds of the flowfield, or, if \code{dtest} is specified, when a
trail becomes too close to an existing point.}

\item{dtest}{The test distance for terminating a trail when it becomes too
close to an existing point. Larger values result in sparser trails. A value
of zero means lines will not be terminated.}
}
\value{
A data frame.
}
\description{
Generate flowfield trails
}
\examples{
make_flowfield() |> make_trails() |> draw_trails()
}
