% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trails.R
\name{make_trails}
\alias{make_trails}
\title{Generate flowfield trails}
\usage{
make_trails(
  flowfield,
  particles = particles_poisson(100, lims(flowfield)),
  max_steps = 1,
  step_length = 0.01,
  direction = c("both", "forward", "backward"),
  dtest = 0
)
}
\arguments{
\item{flowfield}{A data frame containing three columns, x, y, and a: the
coordinates of the flowfield grid and the angle associated with each grid
reference.}

\item{particles}{A data frame containing two columns, x and y: the
coordinates of the starting points (particles) from which trails will be
generated on the flowfield.}

\item{max_steps}{The maximum number of steps to take from each trail seed
point. Trails will be terminated prior to taking this many steps if the go
out of bounds of the flowfield, or, if \code{dtest} is specified, when a
trail becomes too close to an existing point.}

\item{step_length}{The length of each 'step' taken as a trail is generated.
Smaller values result in smoother lines.}

\item{direction}{Which direction should a trail be drawn from a seedpoint?
One of 'forward', 'backward' or 'both'.}

\item{dtest}{The test distance for terminating a trail when it becomes too
close to an existing point. Larger values result in sparser trails. A value
of zero means lines will not be terminated.}
}
\description{
Generate flowfield trails
}
\examples{
make_flowfield() |> make_trails() |> draw_trails()
}
