% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R, R/utils.R
\name{more}
\alias{more}
\alias{more}
\title{more}
\usage{
more(x, grep = NULL, pipe = FALSE)

more(x, grep = NULL, pipe = FALSE)
}
\arguments{
\item{x}{vector of iles}

\item{grep}{string to grep in files (=NULL)}

\item{pipe}{boolean grep piped input 'x' (default = NULL)}

\item{x}{vector of files}

\item{grep}{string to grep in files (default = NULL)}
}
\description{
"more" +/- grep vector of files

"more" +/- grep vector of files
}
\author{
Marcin Imielinski

Marcin Imielinski
}
