% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{purge}
\alias{purge}
\title{purges all output / run directories associated with this Job object (be careful)}
\usage{
purge(object, ...)

\S4method{purge}{Job}(object, check.inputs = TRUE)
}
\description{
purges all output / run directories associated with this Job object (be careful)
}
\author{
Marcin Imielinski
}

