% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{queue}
\alias{queue}
\title{Gets queue associated with the jobs in the Job object}
\usage{
\S4method{queue}{Job}(.Object)
}
\description{
Getting LSF / SGE queue associated with Job object
}
\author{
Marcin Imielinski
}
