% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset.job}
\alias{reset.job}
\title{reset.job

Reset a job with different params}
\usage{
reset.job(
  x,
  ...,
  i = NULL,
  rootdir = x@rootdir,
  jb.mem = x@runinfo$mem,
  jb.cores = x@runinfo$cores,
  jb.time = x@runinfo$time,
  update_cores = 1
)
}
\value{
A Flow job object
}
\description{
reset.job

Reset a job with different params
}
\author{
Kevin Hadi
}
