% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-SearchExtrema.R
\name{SearchExtrema}
\alias{SearchExtrema}
\title{Searches for extrema in a signal.}
\usage{
SearchExtrema(y, which = c("both", "maxima", "minima"))
}
\arguments{
\item{y.}{A numerical vector of signal}

\item{which.}{One of "both", "maxima", "minima". Default is "both"}
}
\value{
A numerical vector of extrema indixes
}
\description{
\code{SearchExtrema} returns extrema of a signal.
}

