% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{sjobs}
\alias{sjobs}
\title{Tracks down any slurm jobs associated with this Job object using a bjobs query (warning can be slow)}
\usage{
\S4method{sjobs}{Job}(.Object)
}
\description{
Tracks down any slurm jobs associated with this Job object using a bjobs query (warning can be slow)
}
\author{
Marcin Imielinski / Kevin Hadi
}
