% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{srun}
\alias{srun}
\title{Runs jobs on Slurm}
\usage{
\S4method{srun}{Job}(.Object, mc.cores = 1, all = FALSE)
}
\arguments{
\item{mc.cores}{Number of parallel cores to run jobs with (=1)}
}
\description{
Runs jobs on Slurm
}
\author{
Marcin Imielinski, Zoran Gajic
}
