% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R, R/utils.R
\name{tailf}
\alias{tailf}
\title{tailf}
\usage{
tailf(x, n = NULL, grep = NULL)

tailf(x, n = NULL, grep = NULL)
}
\arguments{
\item{x}{vector of files}

\item{n}{integer specific number of lines in file (default = NULL)}

\item{grep}{string to grep in files (default = NULL)}
}
\description{
"tail -f" +/- grep vector of files

"tail -f" +/- grep vector of files
}
\author{
Marcin Imielinski

Marcin Imielinski
}
