% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{methods}
\name{time}
\alias{time}
\title{Gets max time limit associated with job object}
\usage{
\S4method{time}{Job}(.Object)
}
\description{
max time limit (in hours or days) the job is allowed to run for
i.e. "days-hours" or "hours" format
"3-00" or "72" specifies 3 days/72 hours
}
\author{
Kevin Hadi
}
