% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R, R/utils.R
\name{xml2task}
\alias{xml2task}
\alias{xml2task}
\title{Makes a best attempt to convert a firehose xml task configuration into a Task object or .task file}
\usage{
xml2task(path, module = NULL, out.file = NULL)

xml2task(path, module = NULL, out.file = NULL)
}
\arguments{
\item{path}{string of pathname to firehose xml task}

\item{module}{string name of Module (default = NULL)}

\item{out.file}{string pathname to *.task file (default = NULL)}
}
\description{
Takes a path to an xml file of a FH task configuration, module path, and outputs a Task object or writes
to a .task fil.e

Takes a path to an xml file of a FH task configuration, module path, and outputs a Task object or writes
to a .task file
}
\author{
Marcin Imielinski

Marcin Imielinski
}
